+++ lib/alice.dart
@@ -1,15 +1,18 @@
 import 'dart:io';
-import 'package:alice/core/alice_chopper_response_interceptor.dart';
-import 'package:alice/core/alice_http_adapter.dart';
-import 'package:alice/model/alice_http_call.dart';
 
-import 'package:chopper/chopper.dart';
-import 'package:http/http.dart' as http;
+import 'package:alice/core/alice_chopper_response_interceptor.dart';
 import 'package:alice/core/alice_core.dart';
 import 'package:alice/core/alice_dio_interceptor.dart';
+import 'package:alice/core/alice_http_adapter.dart';
 import 'package:alice/core/alice_http_client_adapter.dart';
+import 'package:alice/model/alice_http_call.dart';
+import 'package:alice/model/alice_log.dart';
+import 'package:chopper/chopper.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
+import 'package:http/http.dart' as http;
+
+export 'package:alice/model/alice_log.dart';
 
 class Alice {
   /// Should user be notified with notification if there's new request catched
@@ -20,9 +23,6 @@ class Alice {
   /// with sensors)
   final bool showInspectorOnShake;
 
-  /// Should inspector use dark theme
-  final bool darkTheme;
-
   /// Icon url for notification
   final String notificationIcon;
 
@@ -30,8 +30,12 @@ class Alice {
   ///method queue will be used to remove elements.
   final int maxCallsCount;
 
-  ///Directionality of app. Directionality of the app will be used if set to null.
+  ///Directionality of app. Directionality of the app will be used if set to
+  ///null.
   final TextDirection? directionality;
+
+  ///Flag used to show/hide share button
+  final bool? showShareButton;
 
   GlobalKey<NavigatorState>? _navigatorKey;
   late AliceCore _aliceCore;
@@ -43,20 +47,20 @@ class Alice {
     GlobalKey<NavigatorState>? navigatorKey,
     this.showNotification = true,
     this.showInspectorOnShake = false,
-    this.darkTheme = false,
-    this.notificationIcon = "@mipmap/ic_launcher",
+    this.notificationIcon = '@mipmap/ic_launcher',
     this.maxCallsCount = 1000,
     this.directionality,
+    this.showShareButton = true,
   }) {
     _navigatorKey = navigatorKey ?? GlobalKey<NavigatorState>();
     _aliceCore = AliceCore(
       _navigatorKey,
       showNotification: showNotification,
       showInspectorOnShake: showInspectorOnShake,
-      darkTheme: darkTheme,
       notificationIcon: notificationIcon,
       maxCallsCount: maxCallsCount,
       directionality: directionality,
+      showShareButton: showShareButton,
     );
     _httpClientAdapter = AliceHttpClientAdapter(_aliceCore);
     _httpAdapter = AliceHttpAdapter(_aliceCore);
@@ -85,8 +89,10 @@ class Alice {
 
   /// Handle response from HttpClient
   void onHttpClientResponse(
-      HttpClientResponse response, HttpClientRequest request,
-      {dynamic body}) {
+    HttpClientResponse response,
+    HttpClientRequest request, {
+    dynamic body,
+  }) {
     _httpClientAdapter.onResponse(response, request, body: body);
   }
 
@@ -102,8 +108,8 @@ class Alice {
   }
 
   /// Get chopper interceptor. This should be added to Chopper instance.
-  List<ResponseInterceptor> getChopperInterceptor() {
-    return [AliceChopperInterceptor(_aliceCore)];
+  ResponseInterceptor getChopperInterceptor() {
+    return AliceChopperInterceptor(_aliceCore);
   }
 
   /// Handle generic http call. Can be used to any http client.
@@ -111,5 +117,20 @@ class Alice {
     assert(aliceHttpCall.request != null, "Http call request can't be null");
     assert(aliceHttpCall.response != null, "Http call response can't be null");
     _aliceCore.addCall(aliceHttpCall);
+  }
+
+  /// Adds new log to Alice logger.
+  void addLog(AliceLog log) {
+    _aliceCore.addLog(log);
+  }
+
+  /// Adds list of logs to Alice logger
+  void addLogs(List<AliceLog> logs) {
+    _aliceCore.addLogs(logs);
+  }
+
+  /// Returns flag which determines whether inspector is opened
+  bool isInspectorOpened() {
+    return _aliceCore.isInspectorOpened();
   }
 }