+++ lib/core/alice_http_client_extensions.dart
@@ -5,11 +5,14 @@ import 'dart:io';
 import 'package:alice/alice.dart';
 
 extension AliceHttpClientExtensions on Future<HttpClientRequest> {
-  /// Intercept http client with alice. This extension method provides additional
-  /// helpful method to intercept httpClientResponse.
-  Future<HttpClientResponse> interceptWithAlice(Alice alice,
-      {dynamic body, Map<String, dynamic>? headers}) async {
-    final HttpClientRequest request = await this;
+  /// Intercept http client with alice. This extension method provides
+  /// additional helpful method to intercept httpClientResponse.
+  Future<HttpClientResponse> interceptWithAlice(
+    Alice alice, {
+    dynamic body,
+    Map<String, dynamic>? headers,
+  }) async {
+    final request = await this;
     if (body != null) {
       request.write(body);
     }