+++ lib/ui/page/alice_call_details_screen.dart
@@ -1,23 +1,26 @@
 import 'package:alice/core/alice_core.dart';
 import 'package:alice/helper/alice_save_helper.dart';
 import 'package:alice/model/alice_http_call.dart';
-import 'package:alice/utils/alice_constants.dart';
 import 'package:alice/ui/widget/alice_call_error_widget.dart';
 import 'package:alice/ui/widget/alice_call_overview_widget.dart';
 import 'package:alice/ui/widget/alice_call_request_widget.dart';
 import 'package:alice/ui/widget/alice_call_response_widget.dart';
+import 'package:alice/utils/alice_constants.dart';
+import 'package:alice/utils/alice_theme.dart';
 import 'package:collection/collection.dart' show IterableExtension;
 import 'package:flutter/material.dart';
-import 'package:share/share.dart';
+import 'package:share_plus/share_plus.dart';
 
 class AliceCallDetailsScreen extends StatefulWidget {
   final AliceHttpCall call;
   final AliceCore core;
 
-  const AliceCallDetailsScreen(this.call, this.core);
+  const AliceCallDetailsScreen(this.call, this.core, {super.key});
 
   @override
-  _AliceCallDetailsScreenState createState() => _AliceCallDetailsScreenState();
+  State<StatefulWidget> createState() {
+    return _AliceCallDetailsScreenState();
+  }
 }
 
 class _AliceCallDetailsScreenState extends State<AliceCallDetailsScreen>
@@ -35,15 +38,16 @@ class _AliceCallDetailsScreenState extends State<AliceCallDetailsScreen>
       textDirection: widget.core.directionality ?? Directionality.of(context),
       child: Theme(
         data: ThemeData(
-            brightness: widget.core.brightness,
-            accentColor: AliceConstants.lightRed),
+          colorScheme: AliceTheme.getColorScheme(),
+        ),
         child: StreamBuilder<List<AliceHttpCall>>(
           stream: widget.core.callsSubject,
           initialData: [widget.call],
           builder: (context, callsSnapshot) {
             if (callsSnapshot.hasData) {
-              final AliceHttpCall? call = callsSnapshot.data!.firstWhereOrNull(
-                  (snapshotCall) => snapshotCall.id == widget.call.id);
+              final call = callsSnapshot.data!.firstWhereOrNull(
+                (snapshotCall) => snapshotCall.id == widget.call.id,
+              );
               if (call != null) {
                 return _buildMainWidget();
               } else {
@@ -62,15 +66,22 @@ class _AliceCallDetailsScreenState extends State<AliceCallDetailsScreen>
     return DefaultTabController(
       length: 4,
       child: Scaffold(
-        floatingActionButton: FloatingActionButton(
-          backgroundColor: AliceConstants.lightRed,
-          key: const Key('share_key'),
-          onPressed: () async {
-            Share.share(await _getSharableResponseString(),
-                subject: 'Request Details');
-          },
-          child: const Icon(Icons.share),
-        ),
+        floatingActionButton: widget.core.showShareButton == true
+            ? FloatingActionButton(
+                backgroundColor: AliceConstants.lightRed,
+                key: const Key('share_key'),
+                onPressed: () async {
+                  await Share.share(
+                    await _getSharableResponseString(),
+                    subject: 'Request Details',
+                  );
+                },
+                child: Icon(
+                  Icons.share,
+                  color: AliceConstants.white,
+                ),
+              )
+            : null,
         appBar: AppBar(
           bottom: TabBar(
             indicatorColor: AliceConstants.lightRed,
@@ -86,7 +97,7 @@ class _AliceCallDetailsScreenState extends State<AliceCallDetailsScreen>
   }
 
   Widget _buildErrorWidget() {
-    return const Center(child: Text("Failed to load data"));
+    return const Center(child: Text('Failed to load data'));
   }
 
   Future<String> _getSharableResponseString() async {
@@ -94,25 +105,23 @@ class _AliceCallDetailsScreenState extends State<AliceCallDetailsScreen>
   }
 
   List<Widget> _getTabBars() {
-    final List<Widget> widgets = [];
-    widgets.add(const Tab(icon: Icon(Icons.info_outline), text: "Overview"));
-    widgets.add(const Tab(icon: Icon(Icons.arrow_upward), text: "Request"));
-    widgets.add(const Tab(icon: Icon(Icons.arrow_downward), text: "Response"));
-    widgets.add(
+    return <Widget>[
+      const Tab(icon: Icon(Icons.info_outline), text: 'Overview'),
+      const Tab(icon: Icon(Icons.arrow_upward), text: 'Request'),
+      const Tab(icon: Icon(Icons.arrow_downward), text: 'Response'),
       const Tab(
         icon: Icon(Icons.warning),
-        text: "Error",
+        text: 'Error',
       ),
-    );
-    return widgets;
+    ];
   }
 
   List<Widget> _getTabBarViewList() {
-    final List<Widget> widgets = [];
-    widgets.add(AliceCallOverviewWidget(widget.call));
-    widgets.add(AliceCallRequestWidget(widget.call));
-    widgets.add(AliceCallResponseWidget(widget.call));
-    widgets.add(AliceCallErrorWidget(widget.call));
-    return widgets;
+    return [
+      AliceCallOverviewWidget(widget.call),
+      AliceCallRequestWidget(widget.call),
+      AliceCallResponseWidget(widget.call),
+      AliceCallErrorWidget(widget.call),
+    ];
   }
 }