+++ lib/ui/widget/alice_base_call_details_widget.dart
@@ -12,8 +12,10 @@ abstract class AliceBaseCallDetailsWidgetState<T extends StatefulWidget>
     return Row(
       crossAxisAlignment: CrossAxisAlignment.start,
       children: [
-        SelectableText(name,
-            style: const TextStyle(fontWeight: FontWeight.bold)),
+        SelectableText(
+          name,
+          style: const TextStyle(fontWeight: FontWeight.bold),
+        ),
         const Padding(
           padding: EdgeInsets.only(left: 5),
         ),
@@ -24,7 +26,7 @@ abstract class AliceBaseCallDetailsWidgetState<T extends StatefulWidget>
         ),
         const Padding(
           padding: EdgeInsets.only(bottom: 18),
-        )
+        ),
       ],
     );
   }