+++ lib/utils/shake_detector.dart
@@ -1,8 +1,11 @@
 ///Code from https://github.com/deven98/shake
 ///Seems to be not maintained for almost 2 years... (01.03.2021).
+library;
+
 import 'dart:async';
 import 'dart:math';
-import 'package:sensors/sensors.dart';
+
+import 'package:sensors_plus/sensors_plus.dart';
 
 /// Callback for phone shakes
 typedef PhoneShakeCallback = Null Function();
@@ -25,37 +28,39 @@ class ShakeDetector {
   int mShakeCount = 0;
 
   /// StreamSubscription for Accelerometer events
-  StreamSubscription? streamSubscription;
+  StreamSubscription<dynamic>? streamSubscription;
 
   /// This constructor waits until [startListening] is called
-  ShakeDetector.waitForStart(
-      {this.onPhoneShake,
-      this.shakeThresholdGravity = 2.7,
-      this.shakeSlopTimeMS = 500,
-      this.shakeCountResetTime = 3000});
+  ShakeDetector.waitForStart({
+    this.onPhoneShake,
+    this.shakeThresholdGravity = 2.7,
+    this.shakeSlopTimeMS = 500,
+    this.shakeCountResetTime = 3000,
+  });
 
   /// This constructor automatically calls [startListening] and starts detection and callbacks.\
-  ShakeDetector.autoStart(
-      {this.onPhoneShake,
-      this.shakeThresholdGravity = 2.7,
-      this.shakeSlopTimeMS = 500,
-      this.shakeCountResetTime = 3000}) {
+  ShakeDetector.autoStart({
+    this.onPhoneShake,
+    this.shakeThresholdGravity = 2.7,
+    this.shakeSlopTimeMS = 500,
+    this.shakeCountResetTime = 3000,
+  }) {
     startListening();
   }
 
   /// Starts listening to accelerometer events
   void startListening() {
     streamSubscription = accelerometerEvents.listen((AccelerometerEvent event) {
-      final double x = event.x;
-      final double y = event.y;
-      final double z = event.z;
+      final x = event.x;
+      final y = event.y;
+      final z = event.z;
 
-      final double gX = x / 9.80665;
-      final double gY = y / 9.80665;
-      final double gZ = z / 9.80665;
+      final gX = x / 9.80665;
+      final gY = y / 9.80665;
+      final gZ = z / 9.80665;
 
       // gForce will be close to 1 when there is no movement.
-      final double gForce = sqrt(gX * gX + gY * gY + gZ * gZ);
+      final gForce = sqrt(gX * gX + gY * gY + gZ * gZ);
 
       if (gForce > shakeThresholdGravity) {
         final now = DateTime.now().millisecondsSinceEpoch;